exportar recibo de sueldo

+++ src/base/util/MathUtil.java
@@ -6,6 +6,6 @@ public class MathUtil {
 	    for (int i = 0; i < numbers.length; i++) {
 			sum += numbers[i];
 		}
-	    return (sum / numbers.length);
+	    return (double)sum / numbers.length;
 	}
 }
